requirements.txt
asgiref==3.8.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
idna==3.10
Markdown==3.8.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
yt-dlp==2025.7.21
gunicorn==22.0.0
